requirements.txt
pytorch
numpy
pyrtklib
pymap3d
pandas
tqdm
matplotlib
tasgnss